// src/typings/enum.ts
export const enum ActionType {
  ChatInput,
  MessageContext,
  UserContext,
  StringSelectMenu,
  UserSelectMenu,
  RoleSelectMenu,
  MentionableSelectMenu,
  ChannelSelectMenu,
  SelectMenu,
  Button,
  Autocomplete,
  ModalSubmit,
}

export const enum CustomErrorSeverity {
  LOW = 'LOW',
  MEDIUM = 'MEDIUM',
  HIGH = 'HIGH',
  CRITICAL = 'CRITICAL',
}

export const enum CustomErrorType {
  UNKNOWN_ERROR = 'UNKNOWN_ERROR',

  ACTION_IDENTIFIER_ALREADY_EXISTS = 'ACTION_IDENTIFIER_ALREADY_EXISTS',
  ACTION_NOT_FOUND = 'ACTION_NOT_FOUND',
  ACTION_DISABLED = 'ACTION_DISABLED',
  ACTION_WITHOUT_CALLBACK = 'ACTION_WITHOUT_CALLBACK',

  EVENT_IDENTIFIER_ALREADY_EXISTS = 'EVENT_IDENTIFIER_ALREADY_EXISTS',
  EVENT_NOT_FOUND = 'EVENT_NOT_FOUND',
  EVENT_DISABLED = 'EVENT_DISABLED',
  EVENT_WITHOUT_CALLBACK = 'EVENT_WITHOUT_CALLBACK',

  INTERACTION_TYPE_NOT_SUPPORTED = 'INTERACTION_TYPE_NOT_SUPPORTED',

  FILE_NOT_FOUND = 'FILE_NOT_FOUND',
  INVALID_FILE_TYPE = 'INVALID_FILE_TYPE',
  NO_DEFAULT_EXPORT = 'NO_DEFAULT_EXPORT',
  FOLDER_NOT_FOUND = 'FOLDER_NOT_FOUND',
  INVALID_FOLDER_TYPE = 'INVALID_FOLDER_TYPE',
}

export const enum CustomMessageType {
  UNKNOWN_ERROR = 'UNKNOWN_ERROR',

  ACTION_IDENTIFIER_ALREADY_EXISTS = 'ACTION_IDENTIFIER_ALREADY_EXISTS',
  ACTION_NOT_FOUND = 'ACTION_NOT_FOUND',
  ACTION_DISABLED = 'ACTION_DISABLED',
  ACTION_WITHOUT_CALLBACK = 'ACTION_WITHOUT_CALLBACK',

  EVENT_IDENTIFIER_ALREADY_EXISTS = 'EVENT_IDENTIFIER_ALREADY_EXISTS',
  EVENT_NOT_FOUND = 'EVENT_NOT_FOUND',
  EVENT_DISABLED = 'EVENT_DISABLED',
  EVENT_WITHOUT_CALLBACK = 'EVENT_WITHOUT_CALLBACK',

  INTERACTION_TYPE_NOT_SUPPORTED = 'INTERACTION_TYPE_NOT_SUPPORTED',

  FILE_NOT_FOUND = 'FILE_NOT_FOUND',
  INVALID_FILE_TYPE = 'INVALID_FILE_TYPE',
  NO_DEFAULT_EXPORT = 'NO_DEFAULT_EXPORT',
  FOLDER_NOT_FOUND = 'FOLDER_NOT_FOUND',
  INVALID_FOLDER_TYPE = 'INVALID_FOLDER_TYPE',
}